modal_close-btn
modal__cloxe-btn changes for day/night mode;
validation in modal__footer
