+++ src/js/modal.js
@@ -1,4 +1,5 @@
 import { getMovie, findInWatched, findInQueue } from './tmdb';
+import { modalDark } from './dark_theme'
 
 const refs = {
   movieList: document.querySelector('.films__list'),
@@ -81,17 +82,9 @@ async function getMovieAndUpdateUI(selectedMovie) {
       </div>`;
 
     refs.modalContainer.insertAdjacentHTML('beforeend', modalMarkup);
-    if (localStorage.getItem('theme') === 'dark') {
-      const film_title = document.querySelector('.modal__title');
-      const film_values = document.querySelector('.film-values');
-      const modal_description = document.querySelector('.modal__description');
-      const film_vote = document.querySelector('.film-values__votes--color');
-      console.log(modal_description);
-      modal_description.classList.add('dark');
-      film_values.classList.add('film-values--dark_theme');
-      film_vote.classList.add('film_votes--dark');
-      film_title.classList.add('title_dark');
-    }
+
+    modalDark();
+   
   } catch (e) {
     console.log(e);
   }
