+++ src/partials/modal_footer.html
@@ -51,10 +51,10 @@
       </tr>
     </table>
 
-    <ul class="contacts">
-      <h3 class="contacts__title">Contacts for cooperation</h3>
-      
-      <li class="contacts_item"> <a class="contacts__link" href="tel: 012345677"><svg width="12px" height="12px" class="modal__icon">
+
+ <h3 class="contacts__title">Contacts for cooperation</h3>
+    <ul class="contacts">      
+      <li class="contacts_item"> <a class="contacts__link" href="tel:012345677"><svg width="12px" height="12px" class="modal__icon">
                 <use href="./images/symbol-defs.svg#icon-phone-black"></use>
             </svg> 012345677</a></li>
       <li class="contacts_item"><a class="contacts__link" href="mailto:project7@gmail.com"><svg width="12px" height="12px" class="modal__icon">
